refactor(workspace-index): split index state machine into microcrate
