+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "crates/perl-parser-pest",
     "crates/perl-semantic-analyzer",
     "crates/perl-workspace-index",
+    "crates/perl-workspace-index-state-machine",
     "crates/perl-workspace-index-slo",
     "crates/perl-refactoring",
     "crates/perl-incremental-parsing",
@@ -154,7 +155,8 @@ allow = [
     "perl-symbol-index",
     "perl-uri-classify",
     "perl-uri",
-
+    "perl-workspace-index-state-machine",
+    "perl-workspace-index-slo",
     # Tier 3 — Analysis and indexing
     "perl-workspace-index",
     "perl-semantic-analyzer",
@@ -348,6 +350,7 @@ perl-diagnostics-codes = { path = "crates/perl-diagnostics-codes", version = "0.
 
 # Tier 3: Two-level dependencies
 perl-workspace-index = { path = "crates/perl-workspace-index", version = "0.10.0" }
+perl-workspace-index-state-machine = { path = "crates/perl-workspace-index-state-machine", version = "0.10.0" }
 perl-workspace-index-slo = { path = "crates/perl-workspace-index-slo", version = "0.10.0" }
 perl-incremental-parsing = { path = "crates/perl-incremental-parsing", version = "0.10.0" }
 perl-refactoring = { path = "crates/perl-refactoring", version = "0.10.0" }
