+++ crates/perl-workspace-index-state-machine/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "perl-workspace-index-state-machine"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Index lifecycle state machine primitives for workspace indexing"
+readme = "README.md"
+documentation = "https://docs.rs/perl-workspace-index-state-machine"
+keywords = ["perl", "workspace", "index", "state-machine", "lsp"]
+categories = ["development-tools", "text-editors"]
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[dependencies]
+parking_lot = "0.12.5"
+
+[lints]
+workspace = true
