+++ crates/perl-workspace-index-state-machine/README.md
@@ -1,0 +1,6 @@
+# perl-workspace-index-state-machine
+
+Index lifecycle state machine primitives extracted from `perl-workspace-index`.
+
+This microcrate owns the production-ready state model and transition guards for
+workspace indexing lifecycle management.
